requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
python-dotenv>=1.0.0
anthropic>=0.36.0
cadquery>=2.4.0
pydantic>=2.9.0
psycopg2-binary>=2.9.0
pytest>=7.0.0
websockets>=12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
modal>=0.73.0
